app/requirements.txt
aioredis==1.3.1
amqp==5.0.2
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
black==20.8b1
bs4==0.0.1
CacheControl==0.12.6
cachetools==4.1.1
celery==5.0.3
certifi==2019.11.28
cffi==1.14.4
channels==3.0.2
channels-redis==3.2.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
constantly==15.1.0
cryptography==3.2.1
daphne==3.0.1
Django==3.1.3
django-webpack-loader==0.7.0
djangorestframework==3.12.2
firebase-admin==4.4.0
google-api-core==1.23.0
google-api-python-client==1.8.0
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-cloud-core==1.4.4
google-cloud-firestore==2.0.2
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
graphene==2.1.8
graphene-django==2.9.0
graphql-core==2.3.2
graphql-relay==2.0.1
grpcio==1.34.0
gunicorn==20.0.4
hiredis==1.1.0
httplib2==0.18.1
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isodate==0.6.0
jsonfield==3.1.0
kombu==5.0.2
lxml==4.6.2
msgpack==1.0.0
mypy-extensions==0.4.3
nltk==3.4.5
oauth2client==4.1.3
oauthlib==3.1.0
pathspec==0.8.1
promise==2.3
prompt-toolkit==3.0.8
proto-plus==1.11.0
protobuf==3.14.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.0
pytz==2019.3
redis==3.4.1
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.0
Rx==1.6.1
service-identity==18.1.0
singledispatch==3.4.0.3
six==1.14.0
soupsieve==2.0.1
sqlparse==0.4.1
toml==0.10.2
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.8
vine==5.0.0
wcwidth==0.2.5
zope.interface==5.2.0
